+++ tests/testthat/test-lints.R
@@ -14,7 +14,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
-context("lints")
+
 
 if (dir.exists("../../00_pkg_src")) {
   prefix <- "../../00_pkg_src/DIZtools/"