+++ NEWS.md
@@ -1,12 +1,17 @@
 # DIZtools NEWS
 
-## Unreleased (2022-05-20)
+## Unreleased (2022-08-19)
 
+#### Refactorings
+
+* slightly changed print format output of `get_current_timestamp()`
 #### Others
 
+* updated gha
+* updated gha
 * new dev version
 
-Full set of changes: [`v0.0.5...1106ff6`](https://gitlab.miracum.org/miracum/misc/diztools/compare/v0.0.5...1106ff6)
+Full set of changes: [`v0.0.5...d2bd984`](https://gitlab.miracum.org/miracum/misc/diztools/compare/v0.0.5...d2bd984)
 
 ## v0.0.5 (2022-05-18)